gadget: move BuildPartitionList to install and make it unexported
