+++ gadget/install/partition.go
@@ -20,6 +20,7 @@
 package install
 
 import (
+	"bytes"
 	"fmt"
 	"os/exec"
 	"strconv"
@@ -27,19 +28,33 @@ import (
 	"time"
 
 	"github.com/snapcore/snapd/gadget"
+	"github.com/snapcore/snapd/gadget/quantity"
 	"github.com/snapcore/snapd/logger"
 	"github.com/snapcore/snapd/osutil"
+	"github.com/snapcore/snapd/strutil"
 )
 
 var (
 	ensureNodesExist = ensureNodesExistImpl
 )
 
+var createdPartitionGUID = []string{
+	"0FC63DAF-8483-4772-8E79-3D69D8477DE4", // Linux filesystem data
+	"0657FD6D-A4AB-43C4-84E5-0933C84B4F4F", // Linux swap partition
+}
+
+// creationSupported returns whether we support and expect to create partitions
+// of the given type, it also means we are ready to remove them for re-installation
+// or retried installation if they are appropriately marked with createdPartitionAttr.
+func creationSupported(ptype string) bool {
+	return strutil.ListContains(createdPartitionGUID, strings.ToUpper(ptype))
+}
+
 // createMissingPartitions creates the partitions listed in the laid out volume
 // pv that are missing from the existing device layout, returning a list of
 // structures that have been created.
 func createMissingPartitions(dl *gadget.OnDiskVolume, pv *gadget.LaidOutVolume) ([]gadget.OnDiskStructure, error) {
-	buf, created := gadget.BuildPartitionList(dl, pv)
+	buf, created := buildPartitionList(dl, pv)
 	if len(created) == 0 {
 		return created, nil
 	}
@@ -67,6 +82,103 @@ func createMissingPartitions(dl *gadget.OnDiskVolume, pv *gadget.LaidOutVolume) 
 	}
 
 	return created, nil
+}
+
+// buildPartitionList builds a list of partitions based on the current
+// device contents and gadget structure list, in sfdisk dump format, and
+// returns a partitioning description suitable for sfdisk input and a
+// list of the partitions to be created.
+func buildPartitionList(dl *gadget.OnDiskVolume, pv *gadget.LaidOutVolume) (sfdiskInput *bytes.Buffer, toBeCreated []gadget.OnDiskStructure) {
+	sectorSize := dl.SectorSize
+
+	// Keep track what partitions we already have on disk
+	seen := map[quantity.Offset]bool{}
+	for _, s := range dl.Structure {
+		start := s.StartOffset / quantity.Offset(sectorSize)
+		seen[start] = true
+	}
+
+	// Check if the last partition has a system-data role
+	canExpandData := false
+	if n := len(pv.LaidOutStructure); n > 0 {
+		last := pv.LaidOutStructure[n-1]
+		if last.VolumeStructure.Role == gadget.SystemData {
+			canExpandData = true
+		}
+	}
+
+	// The partition index
+	pIndex := 0
+
+	// Write new partition data in named-fields format
+	buf := &bytes.Buffer{}
+	for _, p := range pv.LaidOutStructure {
+		if !p.IsPartition() {
+			continue
+		}
+
+		pIndex++
+		s := p.VolumeStructure
+
+		// Skip partitions that are already in the volume
+		start := p.StartOffset / quantity.Offset(sectorSize)
+		if seen[start] {
+			continue
+		}
+
+		// Only allow the creation of partitions with known GUIDs
+		// TODO:UC20: also provide a mechanism for MBR (RPi)
+		ptype := partitionType(dl.Schema, p.Type)
+		if dl.Schema == "gpt" && !creationSupported(ptype) {
+			logger.Noticef("cannot create partition with unsupported type %s", ptype)
+			continue
+		}
+
+		// Check if the data partition should be expanded
+		size := s.Size
+		if s.Role == gadget.SystemData && canExpandData && quantity.Size(p.StartOffset)+s.Size < dl.Size {
+			size = dl.Size - quantity.Size(p.StartOffset)
+		}
+
+		// Can we use the index here? Get the largest existing partition number and
+		// build from there could be safer if the disk partitions are not consecutive
+		// (can this actually happen in our images?)
+		node := deviceName(dl.Device, pIndex)
+		fmt.Fprintf(buf, "%s : start=%12d, size=%12d, type=%s, name=%q\n", node,
+			p.StartOffset/quantity.Offset(sectorSize), size/sectorSize, ptype, s.Name)
+
+		toBeCreated = append(toBeCreated, gadget.OnDiskStructure{
+			LaidOutStructure: p,
+			Node:             node,
+			Size:             size,
+		})
+	}
+
+	return buf, toBeCreated
+}
+
+func partitionType(label, ptype string) string {
+	t := strings.Split(ptype, ",")
+	if len(t) < 1 {
+		return ""
+	}
+	if len(t) == 1 {
+		return t[0]
+	}
+	if label == "gpt" {
+		return t[1]
+	}
+	return t[0]
+}
+
+func deviceName(name string, index int) string {
+	if len(name) > 0 {
+		last := name[len(name)-1]
+		if last >= '0' && last <= '9' {
+			return fmt.Sprintf("%sp%d", name, index)
+		}
+	}
+	return fmt.Sprintf("%s%d", name, index)
 }
 
 // removeCreatedPartitions removes partitions added during a previous install.
